refactor: switch event price type from double to decimal for improved precision and add validation.

+++ Dtos/Events/EventFilterDto.cs
@@ -2,8 +2,8 @@ namespace Ping.Dtos.Events;
 
 public record EventFilterDto
 {
-    public double? MinPrice { get; init; }
-    public double? MaxPrice { get; init; }
+    public decimal? MinPrice { get; init; }
+    public decimal? MaxPrice { get; init; }
     public DateTime? FromDate { get; init; }
     public DateTime? ToDate { get; init; }
     public int? GenreId { get; init; }
